app/requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-redis==5.4.0
redis==5.0.4
python-decouple
django==4.2.7
requests==2.31.0
gunicorn
psycopg2-binary
africastalking
django-crontab
celery[redis]